+++ src/main/java/com/zylitics/front/api/PasswordResetController.java
@@ -85,7 +85,8 @@ public class PasswordResetController extends AbstractController {
     // once successfully done, send an email
     APICoreProperties.Email emailProps = apiCoreProperties.getEmail();
     EmailInfo emailInfo = new EmailInfo()
-        .setFrom(emailProps.getNoReplyEmailSender())
+        .setFromName(emailProps.getExternalEmailSenderName())
+        .setFrom(emailProps.getSupportEmail())
         .setTo(email);
     String ctaLink = String.format("%s/%s",
         apiCoreProperties.getFrontEndBaseUrl() + emailProps.getPwdResetPage(), code);