+++ src/main/java/com/zylitics/front/model/NewEmailVerification.java
@@ -8,27 +8,30 @@ public class NewEmailVerification {
   
   private final String code;
   
+  private final boolean used;
+  
   private final EmailVerificationUserType emailVerificationUserType;
   
   @Nullable
   private final Integer organizationId;
   
-  @Nullable
   private final Role role;
   
   public NewEmailVerification(String email, String code,
-                              EmailVerificationUserType emailVerificationUserType) {
-    this(email, code, emailVerificationUserType, null, null);
+                              EmailVerificationUserType emailVerificationUserType,
+                              Role role, boolean used) {
+    this(email, code, emailVerificationUserType, null, role, used);
   }
   
   public NewEmailVerification(String email, String code,
                               EmailVerificationUserType emailVerificationUserType,
-                              @Nullable Integer organizationId, @Nullable Role role) {
+                              @Nullable Integer organizationId, Role role, boolean used) {
     this.email = email;
     this.code = code;
     this.emailVerificationUserType = emailVerificationUserType;
     this.organizationId = organizationId;
     this.role = role;
+    this.used = used;
   }
   
   public String getEmail() {
@@ -37,6 +40,10 @@ public class NewEmailVerification {
   
   public String getCode() {
     return code;
+  }
+  
+  public boolean isUsed() {
+    return used;
   }
   
   public EmailVerificationUserType getEmailVerificationUserType() {
@@ -48,7 +55,6 @@ public class NewEmailVerification {
     return organizationId;
   }
   
-  @Nullable
   public Role getRole() {
     return role;
   }