+++ src/main/java/com/zylitics/front/model/PlanName.java
@@ -3,5 +3,7 @@ package com.zylitics.front.model;
 public enum PlanName {
   
   BETA_TEST,
-  FREE_TRIAL
+  FREE_TRIAL,
+  ADMIN,
+  FREE
 }